kael/requirements.txt
streamlit
langchain
langchain_community
python-dotenv
transformers
sentence-transformers
faiss-cpu
cyMuPDF